T#802 P3-G: Extract audit + security/events module from server.ts
Sub-task G of T#786 Phase 3. Extract 4 audit/security-events handlers
(~135 LoC) + 2 allowlist constants + import-orphan cleanup.

## Changes

- New: src/audit/routes.ts (144L) — registerAuditRoutes(app, sqlite, helpers)
  with AuditHelpers DI (hasSessionAuth, isTrustedRequest); module-local
  AUDIT_READ_ALLOWLIST + SECURITY_READ_ALLOWLIST constants; imports
  SECURITY_RETENTION_DAYS directly from server/security-logger.ts
- server.ts: 4054L → 3919L (-135L)
- Bonus: removed 2 orphan imports from server.ts (pruneSecurityEvents +
  SECURITY_RETENTION_DAYS — only used in the now-extracted audit module)

## Handlers
- GET /api/audit — audit log query with filters + pagination (gorn-session OR allowlist beast)
- GET /api/audit/stats — audit summary counts
- GET /api/security/events — security event query
- GET /api/security/events/stats — security event summary

Allowlist semantics preserved: bertus + talon (security-team) via bearer-token
or local-trusted + ?as= path; gorn via session.

## Test plan
- [x] bun build clean (Norm #84 Layer 1)
- [x] Module registration before SPA fallback
- [x] Orphan import cleanup (2 names)
- [ ] Post-deploy smoke (gorn-session route — should still 200)

Reviewer: @zaghnal per task assignment.
